pyproject.toml
[project]
python = "^3.12"

[tool.poetry]
name = "eps-prescription-status-update-api"
version = "0.0.1-alpha"
description = "EPS Prescription Status Update API"
license = "MIT"
authors = [
    "Matthew Popat <matthew.popat1@nhs.net>",
    "Anthony Brown <anthony.brown8@nhs.net>",
    "Kris Szlapa <kris.szlapa1@nhs.net>",
    "Phil Gee <phil.gee@nhs.net>",
    "Adam Brown <adam.brown41@nhs.net>",
    "Jonathan Welch <jonathan.welch1@nhs.net>",
    "Natasa Fragkou <natasa.fragkou@nhs.net>",
    "Jack Spagnoli <jack.spagnoli1@nhs.net>",
    "Tom Smith <tom.smith24@nhs.net>",
    "Sean Steberis <sean.steberis2@nhs.net>",
    "Rizwan Mehmood <rizwan.mehmood1@nhs.net>",
    "Thomas Merrington <thomas.merrington1@nhs.net>"
]
readme = "README.md"
repository = "https://github.com/NHSDigital/eps-prescription-status-update-api"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"
docopt = "^0.6.2"
jsonpath-rw = "^1.4.0"
semver = "^3.0.2"
gitpython = "^3.1.42"
requests = "^2.31.0"
boto3 = "^1.34.94"
argparse = "^1.4.0"
pre-commit = "^3.5.0"
pytest = "^8.2.0"
cfn-lint = "^0.86.4"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
proxygen-cli = "^2.1.11"

[tool.poetry.scripts]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
cfn-lint = "^0.86.4"
black = "^24.4.2"
flake8 = "^7.0.0"
jinja2 = "^3.1.3"
pip-licenses = "^4.4.0"
click = "^8.1.7"
rich = "^13.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
